pyproject.toml
[tool.poetry]
name = "api_pandas"
version = "0.1.0"
description = ""
authors = ["marius.crisan <marius.crisan.tech@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.26.5"
django-filter = "^23.3"
numpy = {version = "^1.26.0", python="^3.10,<3.13"}
pandas = "^2.1.2"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
